+++ src/main/java/donkey/koin/transaction/donkey_kong_transaction/purchase/trigger/PurchaseTriggerDto.java
@@ -12,10 +12,10 @@ public class PurchaseTriggerDto {
     private String token;
 
     @JsonProperty("limit")
-    private int limit;
+    private double limit;
 
     @JsonProperty("coinAmount")
-    private int coinAmount;
+    private double coinAmount;
 
     @JsonProperty("action")
     private PurchaseTriggerAction action;
